spike: demonstrate using POSTs to send data
Sending data over the WebSocket has been causing a number of problems,
so this is a spike/POC to demonstrate how data could be sent using a
POST request to the server instead. The major advantage is that this
uses a different socket, so if there's any sort of buffering-related
issues (a plausible explanation for the ping/pong issues we see), this
should, and does, resolve that.

This code is missing unit tests, logging, and general sanity, but it
demonstrates how it *could* work